made method to convert days into a List
I thought that at most a calendar would need 35 slots as there's on average 5 weeks in a month.  Apparently, there's a chance for a month to need 6 weeks to fit it all, meaning that at most I would need 42 slots which is way too much.  which is why I need a more dynamic storage.